requirements.txt
alembic==1.4.2
click==7.1.1
entrypoints==0.3
flake8==3.7.9
Flask==1.1.2
Flask-Assets==2.0
Flask-Compress==1.4.0
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-SSLify==0.1.5
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
pycodestyle==2.5.0
pyflakes==2.1.1
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
redis==3.4.1
six==1.14.0
SQLAlchemy==1.3.16
webassets==2.0
Werkzeug==1.0.1
WTForms==2.3.1
